small fixes and moved files around for publication
