+++ CMakeLists.txt
@@ -19,6 +19,7 @@ set(CHPLX_VERSION
 )
 
 set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_FLAGS_INIT "-std=c++20")
 
 if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE
