+++ backend/src/utils.cpp
@@ -7,7 +7,7 @@
  * file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  */
 
-#include <hpx/util.hpp>
+#include "hpx/utils.hpp"
 
 #include <string>
 #include <filesystem>
