+++ frontend/lib/util/git-version.cpp
@@ -1,4 +1,3 @@
 namespace chpl {
-  const char* GIT_SHA = "xxxxxxxxxx";
+  const char* GIT_SHA = "ab869c86";
 }
-
