+++ library/CMakeLists.txt
@@ -18,15 +18,15 @@ set(CHPLX_LIBRARY_VERSION
     ${CHPLX_LIBRARY_MAJOR_VERSION}.${CHPLX_LIBRARY_MINOR_VERSION}.${CHPLX_LIBRARY_PATCH_VERSION}.${CHPLX_LIBRARY_BUILD_VERSION}
 )
 
-if(NOT APPLE)
-  set(CMAKE_CXX_STANDARD 20)
-endif()
+set(CMAKE_CXX_STANDARD 20)
+set(CXX_EXTENSIONS OFF)
+set(CMAKE_CXX_STANDARD_REQUIRED YES)
 
 if(NOT WIN32
    AND ${CMAKE_CXX_COMPILER_ID} STREQUAL "Clang"
    AND NOT APPLE
 )
-  set(CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS} -stdlib=libc++)
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20 -stdlib=libc++ -Ofast")
 endif()
 
 project(chplx_library VERSION ${CHPLX_LIBRARY_VERSION})
